Feature/0.2.1 to dev
feat: add debug logging and improve race skip detection reliability

feat: add Symboli Rudolf Support card in event chooser

perf: reduce sleep delay and tap count in team trials navigation sequence

refactor: optimize tap sequence by clicking directly on previous button

refactor: improve type safety and error handling across core modules

feat: add new support cards events data for Mejiro Ryan, Narita Brian, and others

chore: add .windsurf directory to .gitignore

chore: remove windsurf workflows and codeium ignore files

fix: increase hotkey debounce to 0.8s to prevent hook/poll race conditions

feat: improve race lobby button detection with progressive retries and stabilization logic

+++ core/perception/analyzers/support_type.py
@@ -39,7 +39,7 @@ def _circ_med(h: np.ndarray) -> Optional[float]:
     a = np.angle(mu)
     if a < 0:
         a += 2 * np.pi
-    return a * 90.0 / np.pi  # back to [0..179]
+    return float(a * 90.0 / np.pi)  # back to [0..179]
 
 
 def _circ_dist(h1: float, h2: float) -> float:
